--- app/api/wpPrefix.ts
@@ -1,9 +1,0 @@
-export const wpPrefix = () => {
-  const prefix: string | undefined = process.env.WP_BASE_URL;
-
-  if (prefix) {
-    return prefix;
-  } else {
-    throw new Error('WP_BASE_URL is not provided.');
-  }
-};
