--- app/events/README.md
@@ -1,1 +1,0 @@
-- events ディレクトリには、新歓などイベントの告知関連のファイルを置く
