--- lib/nextPrefix.ts
@@ -1,9 +1,0 @@
-export const nextPrefix = () => {
-  const prefix = process.env.NEXT_PUBLIC_BASE_URL;
-
-  if (!prefix) {
-    throw new Error('NEXT_PUBLIC_BASE_URL is not defined');
-  } else {
-    return prefix;
-  }
-};
